riscv::interrupt
Module machine

Enums

Exception

Standard M-mode RISC-V exceptions

Interrupt

Standard M-mode RISC-V interrupts

Functions

cause

Retrieves the cause of a trap in the current hart (machine mode).

disable

Disables all interrupts in the current hart (machine mode).

enable⚠

Enables all the interrupts in the current hart (machine mode).

free

Execute closure f with interrupts disabled in the current hart (machine mode).

nested⚠

Execute closure f with interrupts enabled in the current hart (machine mode).

try_cause

Retrieves the cause of a trap in the current hart (machine mode).