Trait riscv::paging::FrameAllocatorFor

pub trait FrameAllocatorFor<P: PhysicalAddress> {
    // Required method
    fn alloc(&mut self) -> Option<FrameWith<P>>;
}

A trait for types that can allocate a frame of memory.

Required Methods

fn alloc(&mut self) -> Option<FrameWith<P>>

Allocate a frame of the appropriate size and return it if possible.