riscv::register::mstatus
Struct Mstatus

pub struct Mstatus { /* private fields */ }

mstatus register

Implementations

impl Mstatus

pub fn bits(&self) -> usize

Returns the contents of the register as raw bits

pub fn sie(&self) -> bool

Supervisor Interrupt Enable

pub fn set_sie(&mut self, sie: bool)

Update Supervisor Interrupt Enable

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_sie/clear_sie to directly update the CSR.

pub fn mie(&self) -> bool

Machine Interrupt Enable

pub fn set_mie(&mut self, mie: bool)

Update Machine Interrupt Enable

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mie/clear_mie to directly update the CSR.

pub fn spie(&self) -> bool

Supervisor Previous Interrupt Enable

pub fn set_spie(&mut self, spie: bool)

Update Supervisor Previous Interrupt Enable

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_spie` to directly update the CSR.

pub fn ube(&self) -> Endianness

U-mode non-instruction-fetch memory endianness

pub fn set_ube(&mut self, endianness: Endianness)

Update U-mode non-instruction-fetch memory endianness

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_ube to directly update the CSR.

pub fn mpie(&self) -> bool

Machine Previous Interrupt Enable

pub fn set_mpie(&mut self, mpie: bool)

Update Machine Previous Interrupt Enable

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mpie to directly update the CSR.

pub fn spp(&self) -> SPP

Supervisor Previous Privilege Mode

pub fn set_spp(&mut self, spp: SPP)

Update Supervisor Previous Privilege Mode

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_spp to directly update the CSR.

pub fn mpp(&self) -> MPP

Machine Previous Privilege Mode

pub fn set_mpp(&mut self, mpp: MPP)

Update Machine Previous Privilege Mode

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mpp to directly update the CSR.

pub fn fs(&self) -> FS

Floating-point extension state

Encodes the status of the floating-point unit, including the CSR fcsr and floating-point data registers f0–f31.

pub fn vs(&self) -> VS

Vector extension state

pub fn set_fs(&mut self, fs: FS)

Update Floating-point extension state

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_fs to directly update the CSR.

pub fn set_vs(&mut self, vs: VS)

Update vector extension state

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_vs to directly update the CSR.

pub fn xs(&self) -> XS

Additional extension state

Encodes the status of additional user-mode extensions and associated state.

pub fn set_xs(&mut self, xs: XS)

Update Additional extension state

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself.

pub fn mprv(&self) -> bool

Modify Memory PRiVilege

pub fn set_mprv(&mut self, mprv: bool)

Update Modify Memory PRiVilege

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mprv/clear_mprv to directly update the CSR.

pub fn sum(&self) -> bool

Permit Supervisor User Memory access

pub fn set_sum(&mut self, sum: bool)

Update Permit Supervisor User Memory access

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_sum/clear_sum to directly update the CSR.

pub fn mxr(&self) -> bool

Make eXecutable Readable

pub fn set_mxr(&mut self, mxr: bool)

Update Make eXecutable Readable

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mxr/clear_mxr to directly update the CSR.

pub fn tvm(&self) -> bool

Trap Virtual Memory

If this bit is set, reads or writes to satp CSR or execute sfence.vma instruction when in S-mode will raise an illegal instruction exception.

TVM is hard-wired to 0 when S-mode is not supported.

pub fn set_tvm(&mut self, tvm: bool)

Update Trap Virtual Memory

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_tvm/clear_tvm to directly update the CSR.

pub fn tw(&self) -> bool

Timeout Wait

Indicates that if WFI instruction should be intercepted.

If this bit is set, when WFI is executed in S-mode, and it does not complete within an implementation specific, bounded time limit, the WFI instruction will cause an illegal instruction trap; or could always cause trap then the time limit is zero.

TW is hard-wired to 0 when S-mode is not supported.

pub fn set_tw(&mut self, tw: bool)

Update Timeout Wait

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_tw/clear_tw to directly update the CSR.

pub fn tsr(&self) -> bool

Trap SRET

Indicates that if SRET instruction should be trapped to raise illegal instruction exception.

If S-mode is not supported, TSR bit is hard-wired to 0.

pub fn set_tsr(&mut self, tsr: bool)

Update Trap SRET

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_tsr/clear_tsr to directly update the CSR.

pub fn uxl(&self) -> XLEN

Effective xlen in U-mode (i.e., UXLEN).

In RISCV-32, UXL does not exist, and UXLEN is always XLEN::XLEN32.

pub fn set_uxl(&mut self, uxl: XLEN)

Update Effective xlen in U-mode (i.e., UXLEN).

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself.

pub fn sxl(&self) -> XLEN

Effective xlen in S-mode (i.e., SXLEN).

In RISCV-32, SXL does not exist, and SXLEN is always XLEN::XLEN32.

pub fn set_sxl(&mut self, sxl: XLEN)

Update Effective xlen in S-mode (i.e., SXLEN).

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself.

pub fn sbe(&self) -> Endianness

S-mode non-instruction-fetch memory endianness.

In RISCV-32, this field is read from the crate::register::mstatush register.

pub fn set_sbe(&mut self, endianness: Endianness)

Update S-mode non-instruction-fetch memory endianness

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_sbe to directly update the CSR.

pub fn mbe(&self) -> Endianness

M-mode non-instruction-fetch memory endianness

In RISCV-32, this field is read from the crate::register::mstatush register

pub fn set_mbe(&mut self, endianness: Endianness)

Update M-mode non-instruction-fetch memory endianness

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself. See set_mbe to directly update the CSR.

pub fn sd(&self) -> bool

Whether either the FS field or XS field signals the presence of some dirty state

pub fn set_sd(&mut self, sd: bool)

Update whether either the FS field or XS field signals the presence of some dirty state

Note this updates a previously read Mstatus value, but does not affect the mstatus CSR itself.