riscv
Trait ExternalInterruptNumber

pub unsafe trait ExternalInterruptNumber: InterruptNumber { }

Marker trait for enums of target-specific external interrupt numbers.

External interrupts are interrupts caused by external sources (e.g., GPIO, UART, SPI). External interrupts are not retrieved from the mcause CSR. Instead, RISC-V processors have a single core interrupt for all external interrupts. An additional peripheral (e.g., PLIC) is used to multiplex the external interrupts.

Safety

Each enum variant must represent a valid external interrupt number.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.