Crate riscv

Low level access to RISC-V processors

Minimum Supported Rust Version (MSRV)

This crate is guaranteed to compile on stable Rust 1.42 and up. It might compile with older versions but that may change in any new patch release.

Features

This crate provides:

Access to core registers like mstatus or mcause.

Interrupt manipulation mechanisms.

Wrappers around assembly instructions like WFI.

Modules

addr

asm

Assembly instructions

interrupt

Interrupts

paging

register

RISC-V CSR’s

Macros

use_sv32

use_sv39

use_sv48