Module riscv::paging

Structs

MapperFlush

MapperFlushGPA

MapperFlushGPT

PageTableEntryX32

PageTableEntryX64

PageTableEntryX64Printer

PageTableFlags

Possible flags for a page table entry.

PageTableWith

Rv32PageTableWith

This struct is a two level page table with Mapper trait implemented.

Rv39PageTableWith

This struct is a three level page table with Mapper trait implemented.

Rv48PageTableWith

This struct is a four level page table with Mapper trait implemented.

Enums

FlagUpdateError

An error indicating that an update_flags call failed.

MapToError

This error is returned from map_to and similar methods.

UnmapError

An error indicating that an unmap call failed.

Constants

ENTRY_COUNT

RV32_ENTRY_COUNT

RV32_X4_ENTRY_COUNT

RV64_ENTRY_COUNT

RV64_X4_ENTRY_COUNT

Traits

FrameAllocator

Polyfill for default use cases.

FrameAllocatorFor

A trait for types that can allocate a frame of memory.

FrameDeallocator

FrameDeallocatorFor

A trait for types that can deallocate a frame of memory.

Mapper

MapperExt

MapperFlushable

PTE

PTEIterableSlice

Type Definitions

Entries

Entries32

Entries32X4

Entries64

Entries64X4

PageTable

PageTable32X4

PageTable64X4

PageTableEntry

PageTableX32

PageTableX64

Rv32PageTable

Rv32PageTableGuest

Rv32PageTableX4

Rv39PageTable

Rv39PageTableGuest

Rv39PageTableX4

Rv48PageTable

Rv48PageTableGuest

Rv48PageTableX4