Module riscv::register

RISC-V CSR’s

The following registers are not available on 64-bit implementations.

cycleh

timeh

instreth

hpmcounter<3-31>h

mcycleh

minstreth

mhpmcounter<3-31>h

Modules

cycle

cycle register

cycleh

cycleh register

fcsr

Floating-point control and status register

hpmcounter3

Performance-monitoring counter

hpmcounter3h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter4

Performance-monitoring counter

hpmcounter4h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter5

Performance-monitoring counter

hpmcounter5h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter6

Performance-monitoring counter

hpmcounter6h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter7

Performance-monitoring counter

hpmcounter7h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter8

Performance-monitoring counter

hpmcounter8h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter9

Performance-monitoring counter

hpmcounter9h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter10

Performance-monitoring counter

hpmcounter10h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter11

Performance-monitoring counter

hpmcounter11h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter12

Performance-monitoring counter

hpmcounter12h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter13

Performance-monitoring counter

hpmcounter13h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter14

Performance-monitoring counter

hpmcounter14h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter15

Performance-monitoring counter

hpmcounter15h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter16

Performance-monitoring counter

hpmcounter16h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter17

Performance-monitoring counter

hpmcounter17h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter18

Performance-monitoring counter

hpmcounter18h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter19

Performance-monitoring counter

hpmcounter19h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter20

Performance-monitoring counter

hpmcounter20h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter21

Performance-monitoring counter

hpmcounter21h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter22

Performance-monitoring counter

hpmcounter22h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter23

Performance-monitoring counter

hpmcounter23h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter24

Performance-monitoring counter

hpmcounter24h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter25

Performance-monitoring counter

hpmcounter25h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter26

Performance-monitoring counter

hpmcounter26h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter27

Performance-monitoring counter

hpmcounter27h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter28

Performance-monitoring counter

hpmcounter28h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter29

Performance-monitoring counter

hpmcounter29h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter30

Performance-monitoring counter

hpmcounter30h

Upper 32 bits of performance-monitoring counter (RV32I only)

hpmcounter31

Performance-monitoring counter

hpmcounter31h

Upper 32 bits of performance-monitoring counter (RV32I only)

instret

instret register

instreth

instreth register

marchid

marchid register

mcause

mcause register

mcounteren

mcounteren register

mcycle

mcycle register

mcycleh

mcycleh register

medeleg

medeleg register

mepc

mepc register

mhartid

mhartid register

mhpmcounter3

Machine performance-monitoring counter

mhpmcounter3h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter4

Machine performance-monitoring counter

mhpmcounter4h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter5

Machine performance-monitoring counter

mhpmcounter5h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter6

Machine performance-monitoring counter

mhpmcounter6h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter7

Machine performance-monitoring counter

mhpmcounter7h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter8

Machine performance-monitoring counter

mhpmcounter8h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter9

Machine performance-monitoring counter

mhpmcounter9h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter10

Machine performance-monitoring counter

mhpmcounter10h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter11

Machine performance-monitoring counter

mhpmcounter11h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter12

Machine performance-monitoring counter

mhpmcounter12h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter13

Machine performance-monitoring counter

mhpmcounter13h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter14

Machine performance-monitoring counter

mhpmcounter14h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter15

Machine performance-monitoring counter

mhpmcounter15h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter16

Machine performance-monitoring counter

mhpmcounter16h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter17

Machine performance-monitoring counter

mhpmcounter17h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter18

Machine performance-monitoring counter

mhpmcounter18h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter19

Machine performance-monitoring counter

mhpmcounter19h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter20

Machine performance-monitoring counter

mhpmcounter20h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter21

Machine performance-monitoring counter

mhpmcounter21h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter22

Machine performance-monitoring counter

mhpmcounter22h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter23

Machine performance-monitoring counter

mhpmcounter23h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter24

Machine performance-monitoring counter

mhpmcounter24h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter25

Machine performance-monitoring counter

mhpmcounter25h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter26

Machine performance-monitoring counter

mhpmcounter26h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter27

Machine performance-monitoring counter

mhpmcounter27h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter28

Machine performance-monitoring counter

mhpmcounter28h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter29

Machine performance-monitoring counter

mhpmcounter29h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter30

Machine performance-monitoring counter

mhpmcounter30h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmcounter31

Machine performance-monitoring counter

mhpmcounter31h

Upper 32 bits of machine performance-monitoring counter (RV32I only)

mhpmevent3

Machine performance-monitoring event selector

mhpmevent4

Machine performance-monitoring event selector

mhpmevent5

Machine performance-monitoring event selector

mhpmevent6

Machine performance-monitoring event selector

mhpmevent7

Machine performance-monitoring event selector

mhpmevent8

Machine performance-monitoring event selector

mhpmevent9

Machine performance-monitoring event selector

mhpmevent10

Machine performance-monitoring event selector

mhpmevent11

Machine performance-monitoring event selector

mhpmevent12

Machine performance-monitoring event selector

mhpmevent13

Machine performance-monitoring event selector

mhpmevent14

Machine performance-monitoring event selector

mhpmevent15

Machine performance-monitoring event selector

mhpmevent16

Machine performance-monitoring event selector

mhpmevent17

Machine performance-monitoring event selector

mhpmevent18

Machine performance-monitoring event selector

mhpmevent19

Machine performance-monitoring event selector

mhpmevent20

Machine performance-monitoring event selector

mhpmevent21

Machine performance-monitoring event selector

mhpmevent22

Machine performance-monitoring event selector

mhpmevent23

Machine performance-monitoring event selector

mhpmevent24

Machine performance-monitoring event selector

mhpmevent25

Machine performance-monitoring event selector

mhpmevent26

Machine performance-monitoring event selector

mhpmevent27

Machine performance-monitoring event selector

mhpmevent28

Machine performance-monitoring event selector

mhpmevent29

Machine performance-monitoring event selector

mhpmevent30

Machine performance-monitoring event selector

mhpmevent31

Machine performance-monitoring event selector

mideleg

mideleg register

mie

mie register

mimpid

mimpid register

minstret

minstret register

minstreth

minstreth register

mip

mip register

misa

misa register

mscratch

mscratch register

mstatus

mstatus register

mtval

mtval register

mtvec

mtvec register

mvendorid

mvendorid register

pmpaddr0

Physical memory protection address register

pmpaddr1

Physical memory protection address register

pmpaddr2

Physical memory protection address register

pmpaddr3

Physical memory protection address register

pmpaddr4

Physical memory protection address register

pmpaddr5

Physical memory protection address register

pmpaddr6

Physical memory protection address register

pmpaddr7

Physical memory protection address register

pmpaddr8

Physical memory protection address register

pmpaddr9

Physical memory protection address register

pmpaddr10

Physical memory protection address register

pmpaddr11

Physical memory protection address register

pmpaddr12

Physical memory protection address register

pmpaddr13

Physical memory protection address register

pmpaddr14

Physical memory protection address register

pmpaddr15

Physical memory protection address register

pmpcfg0

Physical memory protection configuration pmpcfg0 struct contains pmp0cfg - pmp3cfg for RV32, and pmp0cfg - pmp7cfg for RV64

pmpcfg2

Physical memory protection configuration pmpcfg2 struct contains pmp8cfg - pmp11cfg for RV32, or pmp8cfg - pmp15cfg for RV64

satp

satp register

scause

scause register

scounteren

scounteren register

sepc

sepc register

sie

sie register

sip

sip register

sscratch

sscratch register

sstatus

sstatus register

stval

stval register

stvec

stvec register

time

time register

timeh

timeh register

ucause

ucause register

uepc

uepc register

uie

uie register

uip

uip register

uscratch

uscratch register

ustatus

ustatus register

utval

utval register

utvec

stvec register

Structs

Pmp

Pmp struct holds a high-level representation of a single pmp configuration

Pmpcsr

Enums

Permission

Permission enum contains all possible permission modes for pmp registers

Range

Range enum contains all possible addressing modes for pmp registers