Struct riscv::register::mstatus::Mstatus

pub struct Mstatus { /* private fields */ }

mstatus register

Implementations

impl Mstatus

pub fn sie(&self) -> bool

Supervisor Interrupt Enable

pub fn mie(&self) -> bool

Machine Interrupt Enable

pub fn spie(&self) -> bool

Supervisor Previous Interrupt Enable

pub fn ube(&self) -> Endianness

U-mode non-instruction-fetch memory endianness

pub fn mpie(&self) -> bool

Machine Previous Interrupt Enable

pub fn spp(&self) -> SPP

Supervisor Previous Privilege Mode

pub fn mpp(&self) -> MPP

Machine Previous Privilege Mode

pub fn fs(&self) -> FS

Floating-point extension state

Encodes the status of the floating-point unit, including the CSR fcsr and floating-point data registers f0–f31.

pub fn xs(&self) -> XS

Additional extension state

Encodes the status of additional user-mode extensions and associated state.

pub fn mprv(&self) -> bool

Modify Memory PRiVilege

pub fn sum(&self) -> bool

Permit Supervisor User Memory access

pub fn mxr(&self) -> bool

Make eXecutable Readable

pub fn tvm(&self) -> bool

Trap Virtual Memory

If this bit is set, reads or writes to satp CSR or execute sfence.vma instruction when in S-mode will raise an illegal instruction exception.

TVM is hard-wired to 0 when S-mode is not supported.

pub fn tw(&self) -> bool

Timeout Wait

Indicates that if WFI instruction should be intercepted.

If this bit is set, when WFI is executed in S-mode, and it does not complete within an implementation specific, bounded time limit, the WFI instruction will cause an illegal instruction trap; or could always cause trap then the time limit is zero.

TW is hard-wired to 0 when S-mode is not supported.

pub fn tsr(&self) -> bool

Trap SRET

Indicates that if SRET instruction should be trapped to raise illegal instruction exception.

If S-mode is not supported, TSR bit is hard-wired to 0.

pub fn uxl(&self) -> XLEN

Effective xlen in U-mode (i.e., UXLEN).

In RISCV-32, UXL does not exist, and UXLEN is always XLEN::XLEN32.

pub fn sxl(&self) -> XLEN

Effective xlen in S-mode (i.e., SXLEN).

In RISCV-32, SXL does not exist, and SXLEN is always XLEN::XLEN32.

pub fn sbe(&self) -> Endianness

S-mode non-instruction-fetch memory endianness.

In RISCV-32, this field is read from the crate::register::mstatush register.

pub fn mbe(&self) -> Endianness

M-mode non-instruction-fetch memory endianness

In RISCV-32, this field is read from the crate::register::mstatush register

pub fn sd(&self) -> bool

Whether either the FS field or XS field signals the presence of some dirty state